Correction to use base 2 logarithm for slack vars
The basis values for representing the slack variables ("y") in the
weight inequality constraint should be constructed using the base 2
logarithm as opposed to the natural log.

For example, with a max weight of 50, the previous basis values were:
[1, 2, 4, 43.0]
which cannot represent all possible slack values, such as 8, 9, 10,
etc.  With this fix, the basis values for a max weight of 50 are:
[1, 2, 4, 8, 16, 19.0]