Fix determination of number of slack variables
Following Lucas's paper, Section 2.4, the number of slack variables
should be based on floor(log2(N)) instead of ceiling.  This change
will only impact cases where the max weight is a power of 2.  For
example, with a max weight of 8, the previous basis values were:
[1, 2, 5.0]
which cannot represent the value 4.  With this correction, a max
weight of 8 produces these basis values:
[1, 2, 4, 1.0]