Add very_small.csv example data
This data set illustrates the prior issue with inequality handling:
prior to the log2 fix, the BQM did not correctly define the optimal
solution for this problem.  Previously, for a problem defined with a
max weight of 10, the BQM penalized solutions with weight=5 because
that value was not representable using the selected basis values.
This meant that the BQM did not assign minimum energy to the optimal
solution for this problem, which is to select the item having
weight=5.