Simplify test_knapsack.py
Previously there were two separate tests doing basically the same
thing.  This required running the problem twice unnecessarily.  Updated
so that the problem is only run once, the output is stored, and then
two tests are used to check different aspects of the output.

Also dropping the tests that look for "error" and "warning" in the
output string, which seems unnecessary.