Refactor code and improve screen output
Refactor the solution code into a function solve_knapsack and improve
the screen output.  Instead of just printing "found solution x", print
out more descriptive information specifying the indices of the
selected items, their individual weights (and the total), and the
individual costs (and total).  Update the test as well to handle the
new output format.